Discard most test case usages for now
As most tests require unique test cases, the overhead of using the
framework is dropped for these. To show possible usage, a single
test_case is kept.
